A farm game made on blender game engine
A game depicts scenario of a farm where you have to pick crate of a particular fruit and place it in the designated section of a truck. The objective of the game is to make the truck reach the market with all the fruit crates loaded before the time runs out.